requirements.txt
# 核心库
pytest==8.0.0
pytest-html==4.1.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0
requests==2.31.0
pyyaml==6.0.1

# 数据处理
pandas==2.1.0
openpyxl==3.1.2

# 报告
allure-pytest==2.13.2

# 工具库
python-dotenv==1.0.0
loguru==0.7.2
Faker==20.1.0